crypto.py - Graphic interface added
Update now includes a GUI that allows the user to select from multiple options when live tracking the price of the crypto. Select one or more and it will update in real time. The refresh rate is now improved to every 5000mm/s. To use simply enter the api key after running the code